Change `recordingTimeLeftInSeconds` to `float`

+++ Source/Processors/RecordNode/RecordNodeEditor.cpp
@@ -576,19 +576,20 @@ void FifoMonitor::timerCallback()
 				lastUpdateTime = currentTime;
 				lastFreeSpace = bytesFree;
 
-				recordingTimeLeftInSeconds = (int) (bytesFree / dataRate / 1000.0f);
+				recordingTimeLeftInSeconds = bytesFree / dataRate / 1000.0f;
 
 				LOGD("Data rate: ", dataRate, " bytes/ms");
 
 				// Stop recording and show warning when less than 5 minutes of disk space left
-				if (dataRate > 0 && recordingTimeLeftInSeconds < 60*5) {
+				if (dataRate > 0.0f && recordingTimeLeftInSeconds < (60.0f * 5.0f)) 
+				{
 					CoreServices::setRecordingStatus(false);
 					String msg = "Recording stopped. Less than 5 minutes of disk space remaining.";
 					AlertWindow::showMessageBoxAsync(AlertWindow::WarningIcon, "WARNING", msg);
 				}
 
 				String msg = String(bytesFree / pow(2, 30)) + " GB available\n";
-				msg += String(recordingTimeLeftInSeconds / 60) + " minutes remaining\n";
+				msg += String(int(recordingTimeLeftInSeconds / 60.0f)) + " minutes remaining\n";
 				msg += "Data rate: " + String(dataRate * 1000 / pow(2, 20), 2) + " MB/s";
 				setTooltip(msg);
 			}
